refactor(client): reflect error handling changes
